refactor: 엔티티 일관성 개선 및 변수명 수정 - pk를 "id"로 통일 - @Builder 어노테이션 적용 - 변수명 수정 및 어노테이션 옵션 정리
