--- backend/src/main/java/com/spaceme/User/Domain/AlienType.java
@@ -1,4 +1,0 @@
-package com.spaceme.User.Domain;
-
-public enum AlienType {
-}
