+++ backend/src/main/java/com/spaceme/common/Status.java
@@ -1,4 +1,4 @@
-package com.spaceme.mission.domain;
+package com.spaceme.common;
 
 public enum Status {
     SOON, FAILED, CLEAR, ON_PROGRESS
