+++ backend/src/main/java/com/spaceme/planet/domain/PlanetTheme.java
@@ -3,24 +3,17 @@ package com.spaceme.planet.domain;
 import com.spaceme.galaxy.domain.GalaxyTheme;
 import jakarta.persistence.*;
 import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
 
 @Entity
 @Getter
-@Setter
 public class PlanetTheme {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "planet_theme_id")
-    private Long planetThemeId;
+    private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "galaxy_theme_id")
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
     private GalaxyTheme galaxyTheme;
 
-
+    @Column(nullable = false)
     private String theme;
-
 }
